+++ Assets/Scripts/Massful.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Massful : MonoBehaviour
 {
-    public double Mass => GetComponent<Rigidbody>().mass;
+    public float Mass => GetComponent<Rigidbody>().mass;
     public bool CanBePulled;
 }
